move export print to another file
