+++ srcs/builtin_export.c
@@ -1,40 +1,7 @@
 #include "builtin.h"
 #include "expander.h"
 
-void	print_env_node(t_env_node *node)
-{
-	char	*escaped_value;
-
-	ft_putstr_fd("declare -x ", STDOUT_FILENO);
-	ft_putstr_fd(node->name, STDOUT_FILENO);
-	if (node->value)
-	{
-		escaped_value = get_escaped_str(node->value);
-		ft_putstr_fd("=\"", STDOUT_FILENO);
-		ft_putstr_fd(escaped_value, STDOUT_FILENO);
-		ft_putchar_fd('"', STDOUT_FILENO);
-		free(escaped_value);
-	}
-	ft_putchar_fd('\n', STDOUT_FILENO);
-}
-
-int	export_print(void)
-{
-	t_env_node	*node;
-	t_env_node	*next;
-
-	// TODO: sort environment node list
-	node = g_env->head;
-	while (node)
-	{
-		next = node->next;
-		print_env_node(node);
-		node = next;
-	}
-	return (EXIT_SUCCESS);
-}
-
-char	*get_identifier(char *export_input)
+static char	*get_identifier(char *export_input)
 {
 	char	*delim;
 	char	*identifier;
@@ -51,7 +18,7 @@ char	*get_identifier(char *export_input)
 	return (identifier);
 }
 
-bool	is_identifier_str(char *s)
+static bool	is_identifier_str(char *s)
 {
 	if (s == NULL)
 		return (false);
@@ -66,7 +33,7 @@ bool	is_identifier_str(char *s)
 	return (true);
 }
 
-void	add_exported_env_var(char *export_input)
+static void	add_exported_env_var(char *export_input)
 {
 	char	*delim;
 	char	*name;
@@ -91,7 +58,7 @@ void	add_exported_env_var(char *export_input)
 	free(name);
 }
 
-int	export_set_environs(char **argv)
+static int	export_set_environs(char **argv)
 {
 	int		res;
 	size_t	i;
